fix: error in PastWeather class fix issue with PastWeather class sunshine duration if statement, converting string to int caused an error due to decimal point as discovered when checking weather for 01/01/2020, changed to conversion from string to float

+++ classes.py
@@ -30,7 +30,7 @@ class PastWeather:
         day = date.strftime('%A')
         # Code below to handle singular sunshine hour with correct grammar returned in 
         # instance of 1 hour of sunshine.
-        if int(sunshine_duration) == 1:
+        if float(sunshine_duration) == 1:
             sunshine_string = ["was", "hour"]
 
         os.system('clear')
